Move blocking read
