+++ src/raw/io/output/mod.rs
@@ -23,15 +23,13 @@
 //! our `output` traits.
 
 mod read;
+pub use read::BlockingRead;
+pub use read::BlockingReader;
 pub use read::Read;
 pub use read::ReadExt;
 pub use read::Reader;
 
 pub mod into_reader;
-
-mod blocking_read;
-pub use blocking_read::BlockingRead;
-pub use blocking_read::BlockingReader;
 
 pub mod into_blocking_reader;
